[SW-54] refactor(season): 리뷰 수정 진짜 완료
