+++ src/main/java/com/github/kmu_shell_we/domain/season/controller/AdminSeasonController.java
@@ -4,7 +4,7 @@ import com.github.kmu_shell_we.domain.season.dto.request.CreateSeasonRequest;
 import com.github.kmu_shell_we.domain.season.dto.request.UpdateSeasonRequest;
 import com.github.kmu_shell_we.domain.season.dto.response.SeasonListResponse;
 import com.github.kmu_shell_we.domain.season.dto.response.SeasonResponse;
-import com.github.kmu_shell_we.domain.season.service.SeasonService;
+import com.github.kmu_shell_we.domain.season.service.AdminSeasonService;
 import com.github.kmu_shell_we.global.response.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -23,7 +23,7 @@ import java.util.UUID;
 @Tag(name = "관지라용 시즌 API", description = "시즌 생성, 수정, 삭제를 위한 API")
 public class AdminSeasonController {
 
-    private final SeasonService seasonService;
+    private final AdminSeasonService seasonService;
 
     @GetMapping
     @Operation(summary = "시즌 전체 조회", description = "시즌 목록을 조회합니다.")
