--- src/main/java/com/github/kmu_shell_we/domain/season/repository/AdminSeasonRepository.java
@@ -1,9 +1,0 @@
-package com.github.kmu_shell_we.domain.season.repository;
-
-import com.github.kmu_shell_we.domain.season.entity.Season;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface AdminSeasonRepository extends JpaRepository<Season, UUID> {
-}
